[Go] move Milliseconds into tracing package
You could argue that the Milliseconds type is closely related to the
tracing output format. So move the type into core/tracing and remove
the gtime package.
