+++ go/core/flow.go
@@ -26,7 +26,6 @@ import (
 
 	"github.com/firebase/genkit/go/core/logger"
 	"github.com/firebase/genkit/go/core/tracing"
-	"github.com/firebase/genkit/go/gtime"
 	"github.com/firebase/genkit/go/internal"
 	"github.com/google/uuid"
 	otrace "go.opentelemetry.io/otel/trace"
@@ -172,7 +171,7 @@ type flowState[I, O any] struct {
 	FlowID   string `json:"flowId,omitempty"`
 	FlowName string `json:"name,omitempty"`
 	// start time in milliseconds since the epoch
-	StartTime gtime.Milliseconds `json:"startTime,omitempty"`
+	StartTime tracing.Milliseconds `json:"startTime,omitempty"`
 	Input     I                  `json:"input,omitempty"`
 
 	mu              sync.Mutex
@@ -189,7 +188,7 @@ func newFlowState[I, O any](id, name string, input I) *flowState[I, O] {
 		FlowID:    id,
 		FlowName:  name,
 		Input:     input,
-		StartTime: gtime.ToMilliseconds(time.Now()),
+		StartTime: tracing.ToMilliseconds(time.Now()),
 		Cache:     map[string]json.RawMessage{},
 		Operation: &operation[O]{
 			FlowID: id,
@@ -352,7 +351,7 @@ func (f *Flow[I, O, S]) execute(ctx context.Context, state *flowState[I, O], dis
 	}()
 	ctx = flowContextKey.NewContext(ctx, fctx)
 	exec := &flowExecution{
-		StartTime: gtime.ToMilliseconds(time.Now()),
+		StartTime: tracing.ToMilliseconds(time.Now()),
 	}
 	state.mu.Lock()
 	state.Executions = append(state.Executions, exec)
