+++ go/core/schemas.config
@@ -1,7 +1,7 @@
 # This file holds configuration for the genkit-schema.json file
 # generated by the npm export:schemas script.
 
-core							import github.com/firebase/genkit/go/gtime
+core							import github.com/firebase/genkit/go/core/tracing
 
 # DocumentData type was hand-written.
 DocumentData						omit
@@ -67,8 +67,8 @@ FlowInvokeEnvelopeMessageRunScheduled omit
 Operation						omit
 
 FlowStateExecution				name flowExecution
-FlowStateExecution.startTime	type gtime.Milliseconds
-FlowStateExecution.endTime		type gtime.Milliseconds
+FlowStateExecution.startTime	type tracing.Milliseconds
+FlowStateExecution.endTime		type tracing.Milliseconds
 
 FlowError						name flowError
